Add Redirects
- Add Redirects Controller
- Clean up views, remove unused templates
- Add 301 redirects for all previous URLs

/join-challenger => /
/ (signed-in) => /
/join => /
/about => /
/about/supporter => /supporter
