+++ frontend/app/model/Nav.scala
@@ -17,16 +17,9 @@ object Nav {
 
   val navigation = List(
     NavItem("events", "/events", "Events"),
-    NavItem("whats-on", "/whats-on", "What's On", isPrivate = true,
-      subNav = List(
-        NavItem("calendar", "/whats-on/calendar", "Calendar")
-      )
-    ),
     NavItem("masterclasses", "/masterclasses", "Masterclasses"),
     NavItem("competitions", "/offers-competitions", "Competitions"),
-    NavItem("about", "/about", "About membership"),
-    NavItem("partrons", "/patrons", "Patrons"),
-    NavItem("pricing", "/join", "Pricing"),
+    NavItem("patrons", "/patrons", "Patrons"),
     NavItem("feedback", "/feedback", "Feedback")
   )
