--- frontend/app/views/info/about.scala.html
@@ -1,109 +1,0 @@
-@(pageInfo: model.PageInfo, pageImages: Seq[model.ResponsiveImageGroup])(implicit token: play.filters.csrf.CSRF.Token)
-
-@import configuration.{Links, Videos}
-@import com.gu.membership.salesforce.Tier
-
-@main("About", pageInfo=pageInfo) {
-
-    @* ===== About Membership ===== *@
-    @for(img <- pageImages.find(_.name.contains("intro"))) {
-        <div class="page-slice l-constrained">
-            @fragments.promos.promoPrimary(title="About Membership", img=img, isLead=true) {
-                <div class="text-feature">
-                    <p>Guardian Members are changing the idea of what a news organisation does today.</p>
-                    <p>Join the influential community of journalists, readers and contributors that is the Guardian, and connect to the conversations that matter.</p>
-                </div>
-                <div class="action-group">
-                    <a class="action" href="@routes.Joiner.tierList">Join today</a>
-                    <a class="action action--secondary js-scroll-to" href="#whats-included">
-                        What's included
-                        @fragments.actionIcon("arrow-right")
-                    </a>
-                </div>
-            }
-        </div>
-    }
-
-    @* ===== What is Membership Video ===== *@
-    <div id="what-is-guardian-membership" class="page-slice page-slice--slim l-constrained">
-        @fragments.promos.promoVideoSecondary(
-            "If you believe in Guardian journalism, become a Guardian Member",
-            video=Videos.whatIsMembership,
-            toneClass=Some("tone-trans-brand-dark")){
-            <div class="text-feature">
-                <p>Alan Rusbridger, Polly Toynbee, Jonathan Freedland, Owen Jones and Zoe Williams introduce Guardian Membership.</p>
-            </div>
-        }
-    </div>
-
-    @* ===== Bring the Guardian to life ===== *@
-    @for(img <- pageImages.find(_.name.contains("guardian-live"))) {
-        <div class="page-slice l-constrained">
-            @fragments.promos.promoPrimary(title="Guardian Live: Experience the Guardian brought to life", img=img, stampImageOpt=Some("images/logos/guardian-live-reversed.svg"), toneClass=Some("tone-trans-guardian-live")) {
-                <div class="text-feature">
-                    <p>Book tickets for our programme of discussions, debates, interviews, keynote speeches and festivals.</p>
-                </div>
-                <a class="action" href="@routes.Event.list">See our live events</a>
-            }
-        </div>
-    }
-
-    @for(img <- pageImages.find(_.name.contains("masterclasses"))) {
-        @* ===== Guardian Masterclasses ===== *@
-        <div class="page-slice page-slice--slim l-constrained">
-        @fragments.promos.promoSecondary(Html("Guardian Masterclasses: Learn with the Guardian"), img=img, stampImageOpt = Some("images/logos/guardian-masterclasses-reversed.svg"), toneClass = Some("tone-masterclasses")) {
-            <p class="text-feature">
-                Guardian Masterclasses offer a broad range of short and long courses across a variety of disciplines with a 20% discount for Partners and Patrons.
-            </p>
-            <a class="action" href="@routes.Event.masterclasses">View all courses</a>
-        }
-        </div>
-    }
-
-    @for(img <- pageImages.find(_.name.contains("midland-goods-shed"))) {
-        @* ===== Big ideas ===== *@
-        <div class="page-slice l-constrained">
-        @fragments.promos.promoPrimary("A home for big ideas", img=img, toneClass = Some("tone-guardian-live")) {
-            <div class="text-feature">
-                <p>In 2016, the Guardian will reopen the Midland Goods Shed at London’s King’s Cross to create a new kind of civic space.</p>
-                <p>The building will be a hub for big ideas and stimulating conversations. It will host events,
-                    activities and courses from Guardian Live and institutions we admire, as well as being the home of Guardian Membership.</p>
-            </div>
-            <div class="action-group">
-                <a class="action" target="_blank" href="@Links.membershipBuilding">
-                    More about the building</a>
-                <a class="action action--secondary" target="_blank" href="@Links.membershipBuildingBlog">
-                    Building progress blog
-                    @fragments.actionIcon("arrow-right")
-                </a>
-            </div>
-        }
-        </div>
-    }
-
-    @* ===== Become a Patron ===== *@
-    @for(img <- pageImages.find(_.name.contains("patrons"))) {
-        <div class="page-slice l-constrained">
-            @fragments.promos.promoSecondary(title=Html("Become a Patron and support the Guardian's future"), img=img, toneClass=Some("tone-trans-brand-dark")) {
-                <blockquote class="blockquote blockquote--feature">
-                    The only relationship our journalists have is with our readers. Membership gives the real possibility of deepening the intense bond between the producers and consumers of the Guardian.
-                    <cite class="blockquote__citation">Alan Rusbridger, editor-in-chief of the Guardian</cite>
-                </blockquote>
-                <a class="action" href="@routes.Info.patron">Become a Patron</a>
-            }
-        </div>
-    }
-
-    @* ===== What's Included ===== *@
-    <div class="page-slice tone-tint l-constrained" id="whats-included">
-        <h2 class="page-slice__headline">What's included in Membership</h2>
-        <div class="page-slice__content">
-            @fragments.tier.packageFeature(Tier.Supporter, isReversed=true)
-            @fragments.tier.packageFeature(Tier.Partner, isReversed=true)
-            @fragments.tier.packageFeature(Tier.Patron, isReversed=true)
-            @fragments.tier.friendTeaser()
-        </div>
-    </div>
-
-    @fragments.analytics.videoCampaignPHD.aboutPage()
-}
