--- frontend/app/views/joiner/tierList.scala.html
@@ -1,46 +1,0 @@
-@(pageInfo: model.PageInfo)(implicit token: play.filters.csrf.CSRF.Token)
-
-@import com.gu.membership.salesforce.Tier
-
-@main("Join", pageInfo=pageInfo) {
-
-    <main role="main" class="page-content l-constrained">
-        @fragments.page.pageHeader("Become a member", None)
-
-        <section class="page-section">
-             <div class="page-section__content">
-                <ul class="grid grid--bordered grid--3up-stacked">
-                    <li class="grid__item">
-                        @fragments.tier.packagePromo(Tier.Supporter, Some("#package-benefits-supporter"))
-                    </li>
-                    <li class="grid__item">
-                        @fragments.tier.packagePromo(Tier.Partner, Some("#package-benefits-partner"))
-                    </li>
-                    <li class="grid__item">
-                        @fragments.tier.packagePromo(Tier.Patron, Some("#package-benefits-patron"))
-                    </li>
-                </ul>
-                @fragments.tier.friendTeaser()
-            </div>
-        </section>
-
-        <section class="page-section">
-            <div class="page-section__content">
-                <h3 class="h-section">What's included</h3>
-                <div id="package-benefits-supporter">
-                    @fragments.tier.packageStack(Tier.Supporter)
-                </div>
-                <div id="package-benefits-partner">
-                    @fragments.tier.packageStack(Tier.Partner)
-                </div>
-                <div id="package-benefits-patron">
-                    @fragments.tier.packageStack(Tier.Patron)
-                </div>
-            </div>
-        </section>
-
-        @fragments.analytics.videoCampaignPHD.joinPage()
-
-    </main>
-
-}
