+++ frontend/app/views/patterns/patterns.scala.html
@@ -317,7 +317,7 @@
                     <p> Alan Rusbridger, Polly Toynbee, Jonathan Freedland, Owen Jones and Zoe Williams talk about why readers play such an important part of the Guardian.</p>
                     <p>Sign up as a member, you’ll be supporting free, independent journalism and helping us create original, compelling stories.</p>
                 </div>
-                <a class="action u-no-top-margin" href="@routes.Joiner.tierList">Join today</a>
+                <a class="action u-no-top-margin" href="#">Join today</a>
             }
         </div>
     }
