+++ frontend/conf/routes
@@ -8,19 +8,16 @@ GET            /favicon.ico                       controllers.CacheBustedAssets.
 
 # Home page
 GET            /                                  controllers.FrontPage.index
+GET            /welcome                           controllers.FrontPage.welcome
 
 GET            /gifting                           controllers.Info.giftingPlaceholder
 
-# Healthcheck
+# System
 GET            /healthcheck                       controllers.Healthcheck.healthcheck
-
 GET            /test-users                        controllers.Testing.testUser
 GET            /analytics-off                     controllers.Testing.analyticsOff
 
-GET            /staff/unauthorised                controllers.StaffAuth.unauthorised
-
 # Join
-GET            /join                              controllers.Joiner.tierList
 GET            /join/staff                        controllers.Joiner.staff
 GET            /join/staff/enter-details          controllers.Joiner.enterStaffDetails
 GET            /join/:tier/enter-details          controllers.Joiner.enterDetails(tier: Tier)
@@ -34,11 +31,11 @@ GET            /join/unsupported-browser          controllers.Joiner.unsupported
 
 GET            /choose-tier                       controllers.Joiner.tierChooser
 
-
 # Normal user signin:
 GET            /signin                            controllers.Login.chooseSigninOrRegister(returnUrl: String)
 
 # Staff signin (note, done by OAuth, in addition to regular signin)
+GET            /staff/unauthorised                controllers.StaffAuth.unauthorised
 GET            /staff/login                       controllers.OAuth.login
 GET            /staff/loginAction                 controllers.OAuth.loginAction
 GET            /oauth2callback                    controllers.OAuth.oauth2Callback
@@ -49,7 +46,6 @@ GET            /staff/local-overview              controllers.Staff.localOvervie
 GET            /staff/masterclass-overview        controllers.Staff.masterclassOverview
 GET            /staff/event-details               controllers.Staff.eventDetails
 GET            /staff/admin                       controllers.Staff.admin
-
 
 # Subscription
 POST           /subscription/update-card          controllers.Subscription.updateCard
@@ -95,33 +91,34 @@ POST           /tier/change/:tier                 controllers.TierController.upg
 GET            /tier/change/:tier/summary         controllers.TierController.upgradeThankyou(tier: Tier)
 
 # Information
-
-GET            /about                             controllers.Info.about
-GET            /about/supporter                   controllers.Info.supporter
-GET            /us/supporter                      controllers.Info.supporterUSA
 GET            /patrons                           controllers.Info.patron
+GET            /supporter                         controllers.Info.supporter
+GET            /us/supporter                      controllers.Info.supporterUSA
 GET            /help                              controllers.Info.help
 GET            /feedback                          controllers.Info.feedback
 POST           /feedback                          controllers.Info.submitFeedback
-GET            /join-challenger                   controllers.Info.joinChallenger
-
-# Information - Offers
 GET            /subscriber-offer                  controllers.Info.subscriberOffer
+GET            /offers-competitions               controllers.OffersAndCompetitions.list
 
-# Map static resources from the /public folder to the /assets URL path
-GET            /assets/*assetFile                 controllers.CachedAssets.at(path="/public", assetFile)
+# Styleguide
+GET            /patterns                          controllers.PatternLibrary.patterns
+
 
 GET            /user/me                           controllers.User.me
 GET            /user/me/details                   controllers.User.meDetails
 GET            /user/check-subscriber             controllers.User.checkSubscriberDetails(id: String, postcode: Option[String], lastName: String)
 
+# CSP
+POST           /csp-report                        controllers.Security.cspReport
+
 # Vanity URLS
 GET            /founder                           controllers.VanityUrl.redirect
 
-# Styleguide
-GET            /patterns                          controllers.PatternLibrary.patterns
+# Redirects
+GET            /about                             controllers.Redirects.about
+GET            /about/supporter                   controllers.Redirects.supporter
+GET            /join-challenger                   controllers.Redirects.joinChallenger
+GET            /join                              controllers.Redirects.join
 
-# Offers and Competitions
-GET            /offers-competitions               controllers.OffersAndCompetitions.list
-
-POST           /csp-report                        controllers.Security.cspReport
+# Map static resources from the /public folder to the /assets URL path
+GET            /assets/*assetFile                 controllers.CachedAssets.at(path="/public", assetFile)
